pyproject.toml
[tool.poetry]
name = "sephrasto"
version = "3.5.0"
description = "Ein Charaktergenerator für das DSA-Hausregelsystem Ilaris."
authors = ["Aeolitus", "brzGatsu"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.7"
lxml = "^4.9.3"
PySide6 = "^6.5.1.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
cx-Freeze = "^6.14.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
